requirements.txt
python-telegram-bot
openai
langchain
python-dotenv